requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyarrow>=15.0.0
PyYAML>=6.0.1
joblib>=1.3.2
xgboost>=2.0.0
torch>=2.10.0
shap>=0.45.0
streamlit>=1.34.0
kagglehub>=0.3.8
plotly>=5.20.0
